Move datatable to common
