+++ frontend/src/common/examples/TableExample.tsx
@@ -1,6 +1,6 @@
+import { DataTableCore } from "@/common/datatable/DataTableCore";
 import { PageSections } from "@/common/layout/PageSections";
 import { Tile } from "@/components/composites/Tile";
-import { DataTableCore } from "@/components/datatable/DataTableCore";
 import { Code } from "@/components/ui/code";
 import { CodeBlockData } from "@/components/ui/code-block";
 import { Separator } from "@/components/ui/separator";
