+++ frontend/src/features/parents/components/KidActiveTasksCard.tsx
@@ -1,5 +1,5 @@
-import { DataTableCore } from "@/components/datatable/DataTableCore";
-import { ColumnTitle } from "@/components/datatable/headers/ColumnTitle";
+import { DataTableCore } from "@/common/datatable/DataTableCore";
+import { ColumnTitle } from "@/common/datatable/headers/ColumnTitle";
 import { Spinner } from "@/components/ui/spinner";
 import { useAssignedTasks } from "@/features/tasks/hooks/useAssignedTasks";
 import { ColumnDef } from "@tanstack/react-table";
